Moving logger.py to core/logger.py
